Follow a friendly lead

+++ db/migrate/20170507062401_create_cards.rb
@@ -7,7 +7,7 @@ class CreateCards < ActiveRecord::Migration
       t.string :rank
       t.string :suit
       t.boolean :is_trump, default: false
-      t.integer :strength
+      t.float :strength
 
       t.timestamps null: false
     end
